# api/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Raikada API
  description: API of Raikada, a server and proxy that supports various protocols.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9997
security: []
components:
  schemas:
    OKStatus:
      type: string
      enum: [ok]
    ErrorStatus:
      type: string
      enum: [error]
...
    OK:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/OKStatus'
    Error:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ErrorStatus'
        error:
          type: string
    Info:
      type: object
      properties:
        version:
          type: string
        started:
          type: string
...
    PathConfList:
      type: object
      properties:
        pageCount:
          type: integer
          format: int64
        itemCount:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/PathConf'
...
    PathList:
      type: object
      properties:
        pageCount:
          type: integer
          format: int64
        itemCount:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/Path'
...
    HLSMuxerList:
      type: object
      properties:
        pageCount:
          type: integer
          format: int64
        itemCount:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/HLSMuxer'
...
    RTSPConn:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created:
          type: string
        remoteAddr:
          type: string
        session:
          type: string
          format: uuid
          nullable: true
        tunnel:
          type: string
        inboundBytes:
          type: integer
          format: uint64
        outboundBytes:
          type: integer
          format: uint64
        bytesReceived:
          type: integer
          format: uint64
          deprecated: true
        bytesSent:
          type: integer
          format: uint64
          deprecated: true
    RTSPConnList:
      type: object
      properties:
        pageCount:
          type: integer
          format: int64
        itemCount:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/RTSPConn'
...
paths:
  /v3/info:
    get:
      operationId: info
      tags: [General]
      summary: returns informations about the instance.
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Info'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/auth/jwks/refresh:
    post:
      operationId: authJwksRefresh
      tags: [Authentication]
      summary: Manually refreshes the JWT JWKS.
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/global/get:
    get:
      operationId: configGlobalGet
      tags: [Configuration]
      summary: returns the global configuration.
      description: ''
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalConf'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/global/patch:
    patch:
      operationId: configGlobalSet
      tags: [Configuration]
      summary: patches the global configuration.
      description: all fields are optional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalConf'
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/pathdefaults/get:
    get:
      operationId: configPathDefaultsGet
      tags: [Configuration]
      summary: returns the default path configuration.
      description: ''
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PathConf'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/pathdefaults/patch:
    patch:
      operationId: configPathDefaultsPatch
      tags: [Configuration]
      summary: patches the default path configuration.
      description: all fields are optional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathConf'
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/paths/list:
    get:
      operationId: configPathsList
      tags: [Configuration]
      summary: returns all path configurations.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PathConfList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v3/hlsmuxers/list:
    get:
      operationId: hlsMuxersList
      tags: [HLS]
      summary: returns all HLS muxers.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HLSMuxerList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v3/paths/list:
    get:
      operationId: pathsList
      tags: [Paths]
      summary: returns all paths.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PathList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v3/rtspconns/list:
    get:
      operationId: rtspConnsList
      tags: [RTSP]
      summary: returns all RTSP connections.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPConnList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspconns/get/{id}:
    get:
      operationId: rtspConnsGet
      tags: [RTSP]
      summary: returns a RTSP connection.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPConn'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: connection not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsessions/list:
    get:
      operationId: rtspSessionsList
      tags: [RTSP]
      summary: returns all RTSP sessions.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPSessionList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsessions/get/{id}:
    get:
      operationId: rtspSessionsGet
      tags: [RTSP]
      summary: returns a RTSP session.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPSession'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: session not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsessions/kick/{id}:
    post:
      operationId: rtspSessionsKick
      tags: [RTSP]
      summary: kicks out a RTSP session from the server.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the session.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: session not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsconns/list:
    get:
      operationId: rtspsConnsList
      tags: [RTSP]
      summary: returns all RTSPS connections.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPConnList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsconns/get/{id}:
    get:
      operationId: rtspsConnsGet
      tags: [RTSP]
      summary: returns a RTSPS connection.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPConn'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: connection not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspssessions/list:
    get:
      operationId: rtspsSessionsList
      tags: [RTSP]
      summary: returns all RTSPS sessions.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPSessionList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspssessions/get/{id}:
    get:
      operationId: rtspsSessionsGet
      tags: [RTSP]
      summary: returns a RTSPS session.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPSession'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: session not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspssessions/kick/{id}:
    post:
      operationId: rtspsSessionsKick
      tags: [RTSP]
      summary: kicks out a RTSPS session from the server.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the session.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: session not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtmpconns/list:
    get:
      operationId: rtmpConnsList
      tags: [RTMP]
      summary: returns all RTMP connections.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTMPConnList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtmpconns/get/{id}:
    get:
      operationId: rtmpConnectionsGet
      tags: [RTMP]
      summary: returns a RTMP connection.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTMPConn'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: connection not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtmpconns/kick/{id}:
    post:
      operationId: rtmpConnsKick
      tags: [RTMP]
      summary: kicks out a RTMP connection from the server.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: connection not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtmpsconns/list:
    get:
      operationId: rtmpsConnsList
      tags: [RTMP]
      summary: returns all RTMPS connections.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTMPConnList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtmpsconns/get/{id}:
    get:
      operationId: rtmpsConnectionsGet
      tags: [RTMP]
      summary: returns a RTMPS connection.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTMPConn'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: connection not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtmpsconns/kick/{id}:
    post:
      operationId: rtmpsConnsKick
      tags: [RTMP]
      summary: kicks out a RTMPS connection from the server.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: connection not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/srtconns/list:
    get:
      operationId: srtConnsList
      tags: [SRT]
      summary: returns all SRT connections.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SRTConnList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...